extract models to models.py
