+++ rtcclient/workitem.py
@@ -4,6 +4,7 @@ import xmltodict
 import copy
 from rtcclient import exception
 from requests.exceptions import HTTPError
+from rtcclient.models import Comment
 
 
 class Workitem(FieldBase):
@@ -165,28 +166,3 @@ class Workitem(FieldBase):
 
         self.log.error("No Action named %s", action_name)
         raise exception.NotFound("No Action named %s" % action_name)
-
-
-class Action(FieldBase):
-    log = logging.getLogger("workitem.Action")
-
-    def __str__(self):
-        return self.title
-
-
-class State(FieldBase):
-    log = logging.getLogger("workitem.State")
-
-    def __str__(self):
-        return self.title
-
-
-class Comment(FieldBase):
-    log = logging.getLogger("workitem.Comment")
-
-    def __init__(self, url, rtc_obj, raw_data=None):
-        self.id = url.split("/")[-1]
-        FieldBase.__init__(self, url, rtc_obj, raw_data)
-
-    def __str__(self):
-        return self.id
